카드 추가 구현 부분을 컴포넌트 분리
