+++ src/pages/CardAddPage.tsx
@@ -1,5 +1,7 @@
+import Card from '../components/card-add/Card';
+
 export default function CardAddPage() {
 	return (
-		<></>
+		<Card />
 	);
 }
